+++ R/common-params.R
@@ -1,7 +1,9 @@
 
 
 #' @name common-params
-#' @param id Unique ID string. By default, name of the calling package.
+#' @param id Unique ID string. If this is unset, the name of the calling
+#'   package will be used. Since onetime 0.2.0, not setting `id` is
+#'   `r lifecycle::badge("deprecated")` with a warning.
 #' @param path Directory to store lockfiles.
 #' @param expiry [difftime()] or e.g. [lubridate::duration()] object.
 #'   After this length of time, code will be run again.